# .agents/skills/payluk-api/references/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Payluk ThirdParty API",
    "version": "1.0.0",
    "description": "Merchant / Business API for the Payluk escrow and payments platform.\n\nAll routes are mounted under `/v1` and authenticated with a secret key. Every response uses the standard envelope `{ status, message, data }`.\n\n**Rate limit.** Each secret key may make **10 requests per minute** across all `/v1` routes. Beyond that the API answers `429` with `{ \"status\": 429, \"message\": \"Too many request\" }`. Every response carries `RateLimit` and `RateLimit-Policy` headers (IETF draft-7) so a client can back off before being cut off.\n\n**IP allowlist.** On production, if you have added IP addresses to the allowlist in your dashboard, requests from any other address are refused with `403 Unauthorized IP address`. An empty allowlist admits every address. Staging never checks the caller's IP.\n\n**Merchant super-admin scope.** Escrow, milestone-escrow and category routes accept any approved business key. Every other route (customers, wallet, merchant account, confirming and disputing, additional fee, vault, payments, cards, countries, crypto whitelist) requires the key to belong to a merchant super-admin account and answers `401 Access denied` otherwise.",
    "contact": {
      "name": "Payluk Developer Support",
      "email": "support@payluk.ng",
      "url": "https://payluk.ng"
    }
  },
  "servers": [
    {
      "url": "https://staging.api.payluk.ng",
...
    },
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/v1/escrow/create": {
      "post": {
        "tags": [
          "Escrow"
        ],
        "summary": "Create escrow",
        "operationId": "createEscrow",
        "description": "Generates a standard escrow payment link. Sent as `multipart/form-data` so up to 5 images can be attached under `imageUrl`.\n\nThe **seller** is the account the request acts as: send `customer-id` with the selling customer's id. Omit the header only if your merchant account is itself a business account and is the seller; a merchant that is not a business account gets `401 Merchant cannot create escrow for itself via API`.\n\n`amount`, `purpose`, `whoPays`, `maxDelivery` and `deliveryTimeline` are all required. Amounts are naira with at most two decimal places (kobo), minimum 1000. `maxDelivery` is 1 to 365 `deliveryTimeline` units.",
        "parameters": [
          {
            "$ref": "#/components/parameters/CustomerIdHeaderOptional"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "amount",
                  "purpose",
                  "whoPays",
                  "maxDelivery",
                  "deliveryTimeline"
                ],
                "properties": {
                  "amount": {
                    "type": "number",
                    "minimum": 1000,
                    "description": "Escrow amount in major units (NGN), not kobo: send 150000 for ₦150,000. Kobo are allowed as decimals (₦100.20 is 100.2) and are rounded to two places. Minimum 1000.",
                    "example": 150000
                  },
                  "purpose": {
                    "type": "string",
                    "description": "Short title of what is being sold.",
                    "example": "MacBook Pro 14\""
                  },
                  "description": {
                    "type": "string",
                    "example": "Space grey, sealed"
                  },
                  "whoPays": {
                    "type": "string",
                    "enum": [
                      "buyer",
                      "seller",
                      "both"
                    ],
                    "description": "Which party pays the Payluk fee.",
                    "example": "both"
                  },
                  "maxDelivery": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 365,
                    "description": "Length of the delivery window, in `deliveryTimeline` units. Required. Between 1 and 365.",
                    "example": 3
                  },
                  "deliveryTimeline": {
                    "type": "string",
                    "enum": [
                      "minutes",
                      "hours",
                      "days"
                    ],
                    "example": "days",
                    "description": "Unit for `maxDelivery`. Required."
                  },
                  "totalQuantity": {
                    "type": "integer",
                    "example": 1,
                    "description": "Units for sale off this link. Above 1, each buyer's payment mints its own escrow; see the multi-quantity guide."
                  },
                  "categoryId": {
                    "type": "string",
                    "example": "665f1b2c9a1e4d0012ab3c30"
                  },
                  "imageUrl": {
                    "type": "string",
                    "format": "binary",
                    "description": "Product image. Attach the field up to 5 times to upload multiple images."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Escrow created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EscrowResponse"
                },
                "example": {
                  "status": 201,
                  "message": "Created successfully",
                  "data": {
                    "id": "665f1b2c9a1e4d0012ab3c10",
                    "amount": 150000,
                    "purpose": "MacBook Pro 14\"",
                    "description": "Space grey, sealed",
                    "whoPays": "both",
                    "imageUrl": [
                      "https://cdn.payluk.ng/escrow/abc.png"
                    ],
                    "fee": 3750,
                    "paymentToken": "PY_8AB12C9D3045",
                    "paidAt": null,
                    "status": "PENDING",
                    "state": "AWAITING_PAYMENT",
                    "logs": [],
                    "channel": "API",
                    "isSeller": true,
                    "dispute": null,
                    "paymentDetails": null,
                    "category": null,
                    "completedAt": null,
                    "approvedClaimBy": null,
                    "refundedBy": null,
                    "maxDelivery": 3,
                    "deliveryTimeline": "days",
                    "totalQuantity": 1,
                    "settlementType": "STANDARD",
                    "milestones": [],
                    "createdAt": "2026-06-22T10:15:00.000Z",
                    "updatedAt": "2026-06-22T10:15:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or missing `customer-id` header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "examples": {
                  "validation": {
                    "summary": "Validation error",
                    "value": {
                      "status": 400,
                      "message": "amount is required",
                      "data": {}
                    }
                  },
                  "header": {
                    "summary": "Missing header",
                    "value": {
                      "status": 400,
                      "message": "customer-id header is required",
                      "data": {}
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized: missing token or a merchant attempting to create an escrow for itself.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "examples": {
                  "noToken": {
                    "summary": "No token provided",
                    "value": {
                      "status": 401,
                      "message": "No token provided",
                      "data": {}
                    }
                  },
                  "selfEscrow": {
                    "summary": "Merchant self-escrow",
                    "value": {
                      "status": 401,
                      "message": "Merchant cannot create escrow for itself via API",
                      "data": {}
                    }
                  }
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
    "/v1/customer/permissions/{customerId}": {
      "put": {
        "tags": [
          "Merchant Customers"
        ],
        "summary": "Update customer permissions",
        "operationId": "updateCustomerPermissions",
        "parameters": [
          {
            "$ref": "#/components/parameters/CustomerIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "canWithdraw": {
                    "type": "boolean",
                    "example": true
                  },
                  "canBuy": {
                    "type": "boolean",
                    "example": true
                  },
                  "canSell": {
                    "type": "boolean",
                    "example": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Permissions updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponse"
                },
                "example": {
                  "status": 200,
                  "message": "Customer permissions updated successfully",
                  "data": {
                    "customerId": "6a19b57ffa797d3af379ca2a",
                    "firstname": "firsty",
                    "lastname": "fhgvj",
                    "email": "testdubluber@gmail.com",
                    "phone": "09022331122",
                    "bvn": null,
                    "status": "active",
                    "blockedReason": null,
                    "blockedAt": null,
                    "createdAt": "2026-05-29T15:49:19.848Z",
                    "dob": null,
                    "permissions": {
                      "canWithdraw": true,
                      "canBuy": true,
                      "canSell": false
                    },
                    "countryId": "69b8d4c7f80e1f661b48c9ca"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AccessDenied"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Your secret key as a Bearer token. The key prefix selects the environment: `sk_test_...` (staging) or `sk_live_...` (production); a key on the wrong host is refused with `403 Unauthorized Access`. Each key is limited to 10 requests per minute (`429` beyond that) and, on production, to the IP addresses on your dashboard allowlist when one is configured."
      }
    },
    "parameters": {
...
      "CustomerIdPath": {
        "name": "customerId",
        "in": "path",
        "required": true,
        "description": "The merchant customer's unique identifier.",
        "schema": {
          "type": "string"
        },
        "example": "665f1b2c9a1e4d0012ab3c01"
      },
...
      "CustomerIdHeaderOptional": {
        "name": "customer-id",
        "in": "header",
        "required": false,
        "description": "The merchant customer this request acts on behalf of. Omit it to act as your own business account (only a business-type account can sell on its own behalf).",
        "schema": {
          "type": "string"
        },
        "example": "665f1b2c9a1e4d0012ab3c01"
      }
    },
    "responses": {
...
      "AccessDenied": {
        "description": "Access denied.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "status": 401,
              "message": "Access denied",
              "data": {}
            }
          }
        }
      },
...
      "TooManyRequests": {
        "description": "Rate limit exceeded: more than 10 requests in one minute from this secret key. Read the `RateLimit` header and retry after the window resets.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            },
            "example": {
              "status": 429,
              "message": "Too many request"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "description": "Standard error envelope.",
        "properties": {
          "status": {
            "type": "integer",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Action not allowed"
          },
          "data": {
            "type": "object",
            "example": {}
          }
        }
      },
...
      "EscrowStatus": {
        "type": "string",
        "description": "High-level escrow status.",
        "enum": [
          "PENDING",
          "ONGOING",
          "COMPLETED",
          "REFUNDED",
          "CLAIMED",
          "DISPUTED",
          "INVESTIGATING",
          "SPLIT"
        ]
      },
      "EscrowState": {
        "type": "string",
        "description": "Lifecycle state of the escrow.",
        "enum": [
          "AWAITING_PAYMENT",
          "OPENED",
          "CLOSED"
        ]
      },
      "SettlementType": {
        "type": "string",
        "enum": [
          "STANDARD",
          "MILESTONE",
          "VAULT"
        ]
      },
      "Milestone": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "665f1b2c9a1e4d0012ab3c21"
          },
          "title": {
            "type": "string",
            "example": "Design"
          },
          "description": {
            "type": "string",
            "example": "Figma mockups"
          },
          "amount": {
            "type": "integer",
            "example": 300000
          },
          "dueDate": {
            "type": "string",
            "nullable": true,
            "example": "2026-07-15"
          },
          "customerId": {
            "type": "string",
            "nullable": true,
            "description": "Optional beneficiary. When set, releasing this milestone credits this customer's main balance instead of the seller's.",
            "example": null
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "RELEASED",
              "REFUNDED",
              "SPLIT"
            ],
            "description": "`SPLIT` means this milestone was still held when a dispute was resolved by dividing the funds between both parties.",
            "example": "PENDING"
          },
          "releasedAt": {
            "type": "string",
            "nullable": true,
            "example": null
          }
        }
      },
...
      "VaultParticipant": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "string",
            "example": "665f1b2c9a1e4d0012ab3c40"
          },
          "amount": {
            "type": "integer",
            "example": 600000
          },
          "status": {
            "type": "string",
            "enum": [
              "STAKED",
              "RELEASED",
              "REFUNDED"
            ],
            "example": "STAKED"
          },
          "stakedAt": {
            "type": "string",
            "description": "Unix timestamp (seconds) the stake was locked.",
            "example": "1784900000"
          },
          "settledAt": {
            "type": "string",
            "nullable": true,
            "description": "Unix timestamp (seconds) the stake was released or refunded.",
            "example": null
          }
        }
      },
...
      "Escrow": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "665f1b2c9a1e4d0012ab3c10"
          },
          "amount": {
            "type": "integer",
            "example": 150000
          },
          "purpose": {
            "type": "string",
            "example": "MacBook Pro 14\""
          },
          "description": {
            "type": "string",
            "example": "Space grey, sealed"
          },
          "whoPays": {
            "type": "string",
            "enum": [
              "buyer",
              "seller",
              "both"
            ],
            "example": "both"
          },
          "imageUrl": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "https://cdn.payluk.ng/escrow/abc.png"
            ]
          },
          "fee": {
            "type": "integer",
            "example": 3750
          },
          "additionalFee": {
            "type": "integer",
            "description": "The merchant's own charge on top of the amount (delivery and the like). 0 unless set through Update additional fee.",
            "example": 0
          },
          "additionalFeeRefundable": {
            "type": "boolean",
            "description": "Whether a buyer-favoured dispute refund returns the additional fee to the buyer.",
            "example": true
          },
          "paymentToken": {
            "type": "string",
            "example": "PY_8AB12C9D3045"
          },
          "paidAt": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "status": {
            "$ref": "#/components/schemas/EscrowStatus"
          },
          "state": {
            "$ref": "#/components/schemas/EscrowState"
          },
          "channel": {
            "type": "string",
            "example": "API"
          },
          "isSeller": {
            "type": "boolean",
            "example": true
          },
          "dispute": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "object"
            },
            "example": null
          },
          "category": {
            "type": "object",
            "nullable": true,
            "example": null
          },
          "completedAt": {
            "type": "string",
            "nullable": true,
            "example": null
          },
          "maxDelivery": {
            "type": "integer",
            "example": 3
          },
          "deliveryTimeline": {
            "type": "string",
            "example": "days"
          },
          "totalQuantity": {
            "type": "integer",
            "example": 1
          },
          "settlementType": {
            "$ref": "#/components/schemas/SettlementType"
          },
          "milestones": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Milestone"
            }
          },
          "participants": {
            "type": "array",
            "nullable": true,
            "description": "VAULT escrows only: the customers staking into the pot.",
            "items": {
              "$ref": "#/components/schemas/VaultParticipant"
            }
          },
          "winnerId": {
            "type": "string",
            "nullable": true,
            "description": "VAULT escrows only: the participant declared the winner by the merchant.",
            "example": null
          },
          "split": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EscrowSplit"
              }
            ],
            "nullable": true,
            "description": "Set only when the dispute was resolved with `status: SPLIT`. Null on every other escrow."
          },
          "createdAt": {
            "type": "string",
            "example": "2026-06-22T10:15:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-06-22T10:15:00.000Z"
          }
        }
      },
      "EscrowSplit": {
        "type": "object",
        "description": "How a `SPLIT` dispute resolution divided the funds the escrow was still holding.",
        "properties": {
          "sellerAmount": {
            "type": "number",
            "description": "Released to the seller's main balance.",
            "example": 30000
          },
          "buyerAmount": {
            "type": "number",
            "description": "Returned to the buyer's main balance.",
            "example": 20000
          },
          "pool": {
            "type": "number",
            "description": "What the escrow held at the moment of resolution. Always equals `sellerAmount + buyerAmount`.",
            "example": 50000
          },
          "resolvedAt": {
            "type": "string",
            "description": "Unix seconds, as a string.",
            "example": "1786694400"
          }
        }
      },
      "EscrowResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Operation successful"
          },
          "data": {
            "$ref": "#/components/schemas/Escrow"
          }
        }
      },
...
      "Customer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "665f1b2c9a1e4d0012ab3c01"
          },
          "firstname": {
            "type": "string",
            "example": "Ada"
          },
          "lastname": {
            "type": "string",
            "example": "Eze"
          },
          "email": {
            "type": "string",
            "example": "ada@example.com"
          },
          "phone": {
            "type": "string",
            "example": "08012345678"
          },
          "role": {
            "type": "string",
            "example": "merchant_user"
          },
          "status": {
            "type": "string",
            "example": "active"
          },
          "canBuy": {
            "type": "boolean",
            "example": true
          },
          "canSell": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "CustomerResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "Customer fetched successfully"
          },
          "data": {
            "$ref": "#/components/schemas/Customer"
          }
        }
      },
...
    }
  }
}